+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Changed BTree containers in analysis/use_def to Hash containers.
 * Removed unused regex and lazy_static crates.
+* Add remove_variable function to domain::State